+++ lib/src/models/conversation/get_conversation_response_model.dart
@@ -12,6 +12,7 @@ class GetConversationResponse {
   final Map<String, Conversation>? conversationMeta;
   final Map<String, List<Reaction>>? conversationReactionMeta;
   final Map<String, List<Attachment>>? conversationAttachmentsMeta;
+  final Map<String, List<PollOption>>? conversationPollsMeta;
 
   GetConversationResponse({
     this.chatroomMeta,
@@ -21,6 +22,7 @@ class GetConversationResponse {
     this.conversationMeta,
     this.conversationReactionMeta,
     this.conversationAttachmentsMeta,
+    this.conversationPollsMeta,
   });
 
   factory GetConversationResponse.fromEntity(
@@ -53,6 +55,12 @@ class GetConversationResponse {
           value.map((e) => Attachment.fromEntity(e)).toList(),
         ),
       ),
+      conversationPollsMeta: entity.conversationPollsMeta?.map(
+        (key, value) => MapEntry(
+          key,
+          value.map((e) => PollOption.fromEntity(e)).toList(),
+        ),
+      ),
     );
   }
 
@@ -83,6 +91,12 @@ class GetConversationResponse {
           value.map((e) => e.toEntity()).toList(),
         ),
       ),
+      conversationPollsMeta: conversationPollsMeta?.map(
+        (key, value) => MapEntry(
+          key,
+          value.map((e) => e.toEntity()).toList(),
+        ),
+      ),
     );
   }
 }
@@ -103,6 +117,8 @@ class GetConversationResponseEntity {
   final Map<String, List<ReactionEntity>>? conversationReactionsMeta;
   @JsonKey(name: 'conv_attachments_meta')
   final Map<String, List<AttachmentEntity>>? conversationAttachmentsMeta;
+  @JsonKey(name: 'conv_polls_meta')
+  final Map<String, List<PollOptionEntity>>? conversationPollsMeta;
 
   GetConversationResponseEntity({
     this.chatroomMeta,
@@ -112,6 +128,7 @@ class GetConversationResponseEntity {
     this.conversationReactionsMeta,
     this.conversationAttachmentsMeta,
     this.conversationMeta,
+    this.conversationPollsMeta,
   });
 
   factory GetConversationResponseEntity.fromJson(Map<String, dynamic> json) =>