+++ lib/src/models/poll/add_poll_option_request_model.dart
@@ -1,59 +1,45 @@
-import 'package:likeminds_chat_fl/likeminds_chat_fl.dart';
-
 class AddPollOptionRequest {
-  PollViewData pollViewData;
+  String poll;
   int conversationId;
-  int temporaryId;
 
   AddPollOptionRequest._({
-    required this.pollViewData,
+    required this.poll,
     required this.conversationId,
-    required this.temporaryId,
   });
 
   Map<String, dynamic> toJson() {
     return {
       'conversation_id': conversationId,
       'poll': {
-        'text': pollViewData.text,
+        'text': poll,
       },
     };
   }
 }
 
 class AddPollOptionRequestBuilder {
-  PollViewData? _pollViewData;
+  String? _poll;
   int? _conversationId;
-  int? _temporaryId;
 
-  void pollViewData(PollViewData pollViewData) {
-    _pollViewData = pollViewData;
+  void poll(String poll) {
+    _poll = poll;
   }
 
   void conversationId(int conversationId) {
     _conversationId = conversationId;
   }
 
-  void temporaryId(int temporaryId) {
-    _temporaryId = temporaryId;
-  }
-
   AddPollOptionRequest build() {
-    if (_pollViewData == null) {
-      throw StateError("Poll view data is required");
+    if (_poll == null) {
+      throw StateError("Poll is required");
     }
     if (_conversationId == null) {
       throw StateError("Conversation id is required");
     }
 
-    if (_temporaryId == null) {
-      throw StateError("Temporary id is required");
-    }
-
     return AddPollOptionRequest._(
-      pollViewData: _pollViewData!,
+      poll: _poll!,
       conversationId: _conversationId!,
-      temporaryId: _temporaryId!,
     );
   }
 }