+++ lib/src/models/poll/add_poll_option_response_model.dart
@@ -1,45 +1,38 @@
 import 'package:likeminds_chat_fl/likeminds_chat_fl.dart';
 
 class AddPollOptionResponse {
-  PollViewData? pollViewData;
-  int temporaryId;
+  PollOption? poll;
 
   AddPollOptionResponse({
-    this.pollViewData,
-    required this.temporaryId,
+    this.poll,
   });
 
   factory AddPollOptionResponse.fromEntity(AddPollOptionResponseEntity entity) {
     return AddPollOptionResponse(
-      temporaryId: entity.temporaryId,
-      pollViewData: entity.pollViewData != null
-          ? PollViewData.fromEntity(entity.pollViewData!)
+      poll: entity.pollViewData != null
+          ? PollOption.fromEntity(entity.pollViewData!)
           : null,
     );
   }
 
   AddPollOptionResponseEntity toEntity() {
     return AddPollOptionResponseEntity(
-      temporaryId: temporaryId,
-      pollViewData: pollViewData?.toEntity(),
+      pollViewData: poll?.toEntity(),
     );
   }
 }
 
 class AddPollOptionResponseEntity {
-  PollViewDataEntity? pollViewData;
-  int temporaryId;
+  PollOptionEntity? pollViewData;
 
   AddPollOptionResponseEntity({
     this.pollViewData,
-    required this.temporaryId,
   });
 
   factory AddPollOptionResponseEntity.fromJson(Map<String, dynamic> json) {
     return AddPollOptionResponseEntity(
-      temporaryId: json['temporary_id'],
       pollViewData: json['poll'] != null
-          ? PollViewDataEntity.fromJson(json['poll'] as Map<String, dynamic>)
+          ? PollOptionEntity.fromJson(json['poll'] as Map<String, dynamic>)
           : null,
     );
   }